+++ gcs/CHANGES.md
@@ -3,6 +3,7 @@
 ## Next
 1. PR #1511 - [Feature] Storage Client Caching for downscoped token
 2. Enable move by default.
+3. [Feature] Add support for cloud logging and prefix invocation ID to logs for correlation.
 
 ## 3.1.7 - 2025-09-10
 1. PR #1484 - [Feature] Enabled write checksum by default