+++ gcs/CHANGES.md
@@ -1,5 +1,7 @@
 # Release Notes
 
+## Next
+
 ## 3.0.12 - 2025-09-26
 1. Enable move by default