+++ gcs/CHANGES.md
@@ -1,6 +1,8 @@
 # Release Notes
 
 ## Next
+
+## 3.1.8 - 2025-09-26
 1. PR #1511 - [Feature] Storage Client Caching for downscoped token
 2. Enable move by default.